+++ client/src/pages/Step-2/Step2.jsx
@@ -3,7 +3,6 @@ import DesignsInfo from "../../utils/Designs/DesingsInfo";
 import SampleCard from "../components/Step-2/SampleCard";
 import html2canvas from "html2canvas";
 import jsPDF from "jspdf";
-import axios from "axios";
 import { useRef } from "react";
 import { useLocation } from "react-router-dom";
 import { useEffect, useState } from "react";
@@ -11,9 +10,6 @@ import NB001 from "../Templates_designs/NB001.jsx";
 import NB002 from "../Templates_designs/NB002.jsx";
 import { FileDown } from "lucide-react";
 import toast, { Toaster } from "react-hot-toast";
-import Cookies from "js-cookie";
-
-const baseURL = process.env.REACT_APP_BASE_API_URL;
 
 
 export default function Step2() {
@@ -22,22 +18,15 @@ export default function Step2() {
     billedTo: {},
     items: [],
     owner : {},
+    grandTotal : 0,
   });
   useEffect(() => {
     if (location.state) {
-      const { billedTo, items, owner } = location.state;
-      setInvoiceDetails({ billedTo, items, owner });
+      const { billedTo, items, owner, grandTotal } = location.state;
+      setInvoiceDetails({ billedTo, items, owner, grandTotal });
     }
   }, [location.state]);
 
-  const [grandTotal, setGrandTotal] = useState(0);
-  useEffect(() => {
-    let GT = 0;
-    invoiceDetails.items.forEach((item) => {
-      GT += parseInt(item.amount);
-    });
-    setGrandTotal(GT);
-  }, [invoiceDetails]);
 
   const [invoiceCode, setInvoiceCode] = useState({
     code: "NB001",
@@ -47,13 +36,13 @@ export default function Step2() {
     {
       code: "NB001",
       component: (
-        <NB001 invoiceDetails={invoiceDetails} grandTotal={grandTotal} />
+        <NB001 invoiceDetails={invoiceDetails} />
       ),
     },
     {
       code: "NB002",
       component: (
-        <NB002 invoiceDetails={invoiceDetails} grandTotal={grandTotal} />
+        <NB002 invoiceDetails={invoiceDetails} />
       ),
     },
   ];
@@ -87,37 +76,7 @@ export default function Step2() {
     });
   };
 
-  const invoice = async (e) => {
-    e.preventDefault();
 
-    const token = Cookies.get("nb_token");
-
-    try {
-      const response = await axios.post(`${baseURL}/api/user/addInvoice`, {
-        invoiceNo : "NB0001",
-        billedTo : invoiceDetails["billedTo"].client,
-        amount : grandTotal,
-      },
-      {
-        headers : {
-          'content-type' : 'application/json',
-          Authorization : `Bearer ${token}`,
-        }
-      }
-    );
-
-      if(response.data.success){
-        toast.success("Invoice Stored");
-      }
-      else{
-        toast.error("Failed to add invoice to database.");
-      }
-      
-    } catch (error) {
-      console.log("Error adding invoice : ",error);
-      toast.error("Interal Server error!");
-    }
-  };
 
   return (
     <div className="bg-img-common">
@@ -144,9 +103,10 @@ export default function Step2() {
         </div>
 
         <div className="w-full flex justify-center items-center my-10">
-          <button onClick={(e) => {
+          <button onClick={() => 
+          {
             downloadPDF();
-            invoice(e);
+          toast.success("File Downloaded");
           }} class="flex justify-center items-center space-x-3 relative py-2 px-8 text-black text-base font-semibold nded-full overflow-hidden bg-violet-500 rounded-full transition-all duration-400 ease-in-out shadow-md hover:scale-105 hover:text-white hover:shadow-lg active:scale-90 before:absolute before:top-0 before:-left-full before:w-full before:h-full before:bg-gradient-to-r before:from-blue-500 before:to-blue-300 before:transition-all before:duration-500 before:ease-in-out before:z-[-1] before:rounded-full hover:before:left-0 shadow-slate-900">
             <span>
               <FileDown />
